refactor: PartyAccountCad의 템플릿 파트  CardTemplate컴포넌트로 분리
